Accepts Nullable in recordWholeDayEvent and improve attendance UI
Moved testStudentForm export from dashboard.js to students.js for better modularity and imported it in app.js. Improved time formatting in student_attendance.js and updated attendance table to show formatted times and 'Absent' status.
Fixed PHP method signature to accept nullable StudentAttendance. Minor UI and comment adjustments in events.blade.php.
